% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_to_dataframe.R
\name{api_to_dataframe}
\alias{api_to_dataframe}
\title{Given a url submit the request and return the results as a
dataframe.}
\usage{
api_to_dataframe(url, data = NULL)
}
\arguments{
\item{url}{string}

\item{data}{dataframe}
}
\value{
dataframe
}
\description{
Given a url submit the request and return the results as a
dataframe. If the response is paginated, the functions is called
recursively until all of the pages have been loaded ($next is
null) (this should have a limit on it some day.)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
