% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_query_string.R
\name{build_query_string}
\alias{build_query_string}
\title{Given a list of filters return a url query string}
\usage{
build_query_string(filters)
}
\arguments{
\item{filters}{list}
}
\value{
A formated string that matches the url search parameters
}
\description{
Given a list of filters (key-value pairs), this functions returns
a url query string that can be applied to an api
endpoint. Multiple list elements are separated by a '&' and list
elements with multiple values are returned as comma separated
values.
}
\examples{
# return "?size=sm"
build_query_string(list(size="sm"))
# return "?colour=red,blue"
build_query_string(list(colours=c("red","blue")))
# return "?size=sm&colour=red,blue"
build_query_string(list(size='sm', colours=c("red","blue")))


}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
