% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_filters.R
\name{check_filters}
\alias{check_filters}
\title{Check the filters against api filters}
\usage{
check_filters(endpoint, filters)
}
\arguments{
\item{endpoint}{\itemize{
\item the name of the api endpoint
}}

\item{filters}{\itemize{
\item vector of filter names that are to be applied to the
endpoint.
}}
}
\value{
list
}
\description{
This function is used within data fetching functions to verify that
the list of provided filters are known to exist.  If a filter
cannot be found, a warning is printed out the screen. If the global
api filter object does exist, it is populated before proceeding.
}
\examples{
check_filters("stocking_events", list("year"=2012, "species"="LAT"))
check_filters("stocking_events", list("year"=2012, "species"="LAT", 'col'="red"))
check_filters("foo", list("year"=2012, "species"="LAT"))
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
