% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_agencies.R
\name{get_agencies}
\alias{get_agencies}
\title{Get agencies data from Great Lakes Fish Stocking API}
\usage{
get_agencies()
}
\value{
dataframe
}
\description{
This function is a essentially a wrapper around api_to_dataframe
and calls the api endpoint to retrieve all of the agencies currently
in the Great Lakes Fish Stocking database (www.fsis.glfc.org).
}
\examples{
get_agencies()
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
