% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api_filters.R
\name{get_api_filters}
\alias{get_api_filters}
\title{Create global list of available API filters}
\usage{
get_api_filters()
}
\value{
list
}
\description{
This function connects to the openapi/swagger endpoint provided by
http://fsis.glfc.org/  api and fetches all of the available
filters for each endpoint.  The filters are available in a global
list 'api_filters' which subsequently used by other functions -
check_filters, show_filters.  Generally, this function is not
intended to be called directly by the user.
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
