% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conditions.R
\name{get_conditions}
\alias{get_conditions}
\title{Get the Fish Condition from Great Lakes Fish Stocking API}
\usage{
get_conditions()
}
\value{
dataframe
}
\description{
This function is a wrapper around api_to_dataframe and calls the
api endpoint to retrieve condition codes and their associated
descriptions from the Great Lakes Fish Stocking database
(www.fsis.glfc.org).  The returned dataframe contains the code and
a short description for each one,
}
\examples{

get_conditions()

}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
