% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cwt_events.R
\name{get_cwt_events}
\alias{get_cwt_events}
\title{Get CWT stocking events from Great Lakes Fish Stocking API}
\usage{
get_cwt_events(filter_list = list())
}
\arguments{
\item{filter_list}{list}
}
\value{
dataframe
}
\description{
This function is a wrapper around api_to_dataframe and calls the
api endpoint to retrieve Coded Wire Tag (cwt) stocking events the
Great Lakes Fish Stocking database (www.fsis.glfc.org).  This
function takes an optional filter list which can be used to return
stocking event based on several attributes of either the cwt or
the associated stocking event. If no filter list is provided, only
the first page of results are fetched from the api (and a warning
is issued)
}
\details{
See http://fsis.glfc.com/api/api/public_urls/#stocking-get_cwt_events-list
for the full list of available filter keys (query parameters)
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
