% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grid10s.R
\name{get_grid10s}
\alias{get_grid10s}
\title{Get 10-Minute Grids data from Great Lakes Fish Stocking API}
\usage{
get_grid10s(filter_list = list())
}
\arguments{
\item{filter_list}{list}
}
\value{
dataframe
}
\description{
This function is a wrapper around api_to_dataframe
and calls the api endpoint to retrieve all of the 10-minute grids
currently in the Great Lakes Fish Stocking database
(www.fsis.glfc.org).  This function takes an
optional filter list which can be used to return 10-minute grids
associated with a particular lake, juristiction, or management unit.
}
\details{
See http://fsis.glfc.com/api/public_urls/#v1-common-grid10-list
for the full list of available filter keys (query parameters)
}
\examples{

get_grid10s()
get_grid10s(list(lake="ON"))
get_grid10s(list(mu_type="stat_dist"))

}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
