% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jurisdictions.R
\name{get_jurisdictions}
\alias{get_jurisdictions}
\title{Get jurisdiction data from Great Lakes Fish Stocking API}
\usage{
get_jurisdictions(filter_list = list())
}
\arguments{
\item{filter_list}{list}
}
\value{
dataframe
}
\description{
This function is a wrapper around api_to_dataframe
and calls the api endpoint to retrieve all of the jurisdictions
currently in the Great Lakes Fish Stocking database
(www.fsis.glfc.org).  Jurisdiction is defined as the waters of a
lake within the bounds of a province or state.
"The Michigan waters of Lake Huron".  This function takes an
optional filter list which can be used to return jurisdictions
associated with a particular lake or state/province.
}
\details{
See http://fsis.glfc.org/api/public_urls/#v1-common-jurisdiction-list for
the full list of available filter keys (query parameters)
}
\examples{

get_jurisdictions()
get_jurisdictions(list(lake="HU"))
get_jurisdictions(list(stateprov="MI"))

}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
