% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lakes.R
\name{get_lakes}
\alias{get_lakes}
\title{Get lakes data from Great Lakes Fish Stocking API}
\usage{
get_lakes()
}
\value{
dataframe
}
\description{
This function is  a wrapper around api_to_dataframe
and calls the api endpoint to retrieve all of the lakes currently
in the Great Lakes Fish Stocking database (www.fsis.glfc.org)
}
\examples{
lakes <- get_lakes()
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
