% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lifestages.R
\name{get_lifestages}
\alias{get_lifestages}
\title{Get the Life Stages from Great Lakes Fish Stocking API}
\usage{
get_lifestages()
}
\value{
dataframe
}
\description{
This function is a wrapper around api_to_dataframe and calls the
api endpoint to retrieve all of the life stages currently in the
Great Lakes Fish Stocking database (www.fsis.glfc.org).  The
returned dataframe contains the abbreviation used for the
lifestage and a simple description.
}
\examples{

get_lifestages()

}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
