% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_management_units.R
\name{get_management_units}
\alias{get_management_units}
\title{Get Management Units from Great Lakes Fish Stocking API}
\usage{
get_management_units(filter_list = list())
}
\arguments{
\item{filter_list}{list}
}
\value{
dataframe
}
\description{
This function is a wrapper around api_to_dataframe
and calls the api endpoint to retrieve all of the management units
currently in the Great Lakes Fish Stocking database
(www.fsis.glfc.org).  This function takes an
optional filter list which can be used to return management units
associated with a particular lake, state/province, juristiction, or management
unit type (currently 'stat_dist' or 'qma').
}
\details{
See http://fsis.glfc.com/api/public_urls/#v1-common-management_unit-list
for the full list of
available filter keys (query parameters)
}
\examples{

get_grid10s()
get_grid10s(list(lake="ON"))
get_grid10s(list(mu_type="stat_dist"))

}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
