% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_strains.R
\name{get_raw_strains}
\alias{get_raw_strains}
\title{Get the list raw strains from Great Lakes Fish Stocking API}
\usage{
get_raw_strains(filter_list = list())
}
\arguments{
\item{filter_list}{list}
}
\value{
dataframe
}
\description{
This function is a wrapper around api_to_dataframe and calls the
api endpoint to retrieve all of the raw (detailed) strains
currently in the Great Lakes Fish Stocking database
(www.fsis.glfc.org).  Raw Stains are detailed values submitted by
contributing agencies. The get_raw_strain() function takes an
optional filter list which can be used to return raw strains
associated with a particular species or strain.
}
\details{
See http://fsis.glfc.com/api/docs/#v1-common-raw-strainraw-list for
the full list of available filter keys (query parameters)
}
\examples{

get_raw_strains()
get_raw_strains(list(species="LAT"))
get_raw_strains(list(strain="SEN"))
}
\seealso{
get_strains
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
