% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_root_url.R
\name{get_root_url}
\alias{get_root_url}
\title{Internal function to return the root url for our api.}
\usage{
get_root_url()
}
\value{
string
}
\description{
This function gis intended to be used by all of the functions that
call an api endpoint.  If the api endpoint changes in the future,
(ie - the domain name changes or the version changes) we just need
to update this value.  Note that the url string does \emph{NOT} include
a trailing slash.
}
\details{
To use a local version of fsdviz api during development, set the
local environment variable FSDVIZ_DEV to True:

~Sys.setenv(FSDVIZ_DEV=TRUE)~
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
