% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species.R
\name{get_species}
\alias{get_species}
\title{Get the list species from Great Lakes Fish Stocking API}
\usage{
get_species()
}
\value{
dataframe
}
\description{
This function is a wrapper around api_to_dataframe
and calls the api endpoint to retrieve all of the species
currently in the Great Lakes Fish Stocking database
(www.fsis.glfc.org).  The returned dataframe contains the
abbreviation, common_name, scientific_name, amd species_code fo
each species.
}
\examples{

get_species()

}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
