% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_state_provinces.R
\name{get_state_provinces}
\alias{get_state_provinces}
\title{Get state/province data from Great Lakes Fish Stocking API}
\usage{
get_state_provinces(filter_list = list())
}
\arguments{
\item{filter_list}{list}
}
\value{
dataframe
}
\description{
This function is a essentially a wrapper around api_to_dataframe
and calls the api endpoint to retrieve all of the state_provinces
currently in the Great Lakes Fish Stocking database
(www.fsis.glfc.org).  This function takes an
option filter list which can be used to return state_provinces
associated with a particular country.
}
\examples{

get_state_provinces()
get_state_provinces(list(country="CA"))

}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
