% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stocking_events.R
\name{get_stocking_events}
\alias{get_stocking_events}
\title{Get stocking events from Great Lakes Fish Stocking API}
\usage{
get_stocking_events(filter_list = list())
}
\arguments{
\item{filter_list}{list}
}
\value{
dataframe
}
\description{
This is one of the main functions in the GLFishStockR package.  It
calls the api endpoint to retrieve stocking events
the Great Lakes Fish Stocking database
(www.fsis.glfc.org).  This function takes an
optional filter list which can be used to return stocking event
based on several attributes of the
stocking event such as species, lake, agency, first year, as last year.
}
\details{
See http://fsis.glfc.com/api/public_urls/#v1-stocking-event-list
for the full list of available filter keys (query parameters)
}
\examples{

get_cwt_events(list(lake="ON", species=c('BNT', 'RBT')))

}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
