% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stocking_methods.R
\name{get_stocking_methods}
\alias{get_stocking_methods}
\title{Get the Stocking Methods from the Great Lakes Fish Stocking API}
\usage{
get_stocking_methods()
}
\value{
dataframe
}
\description{
This function is a wrapper around api_to_dataframe and calls the
api endpoint to retrieve all of the stocking methods currently in
the Great Lakes Fish Stocking database (www.fsis.glfc.org).  The
returned dataframe contains the abbreviation used for the stocking
method (stk_meth) and the a longer description.
}
\examples{

get_stocking_methods()

}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
