% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_strains.R
\name{get_strains}
\alias{get_strains}
\title{Get the list strains from Great Lakes Fish Stocking API}
\usage{
get_strains(filter_list = list())
}
\arguments{
\item{filter_list}{list}
}
\value{
dataframe
}
\description{
This function is a wrapper around api_to_dataframe and calls the
api endpoint to retrieve all of the strains currently in the Great
Lakes Fish Stocking database (www.fsis.glfc.org).  Strains is the
high level group used to aggregrate detail strain information
provided by the stocking agecies (see get_raw_strain() to access
those values).  The get_strain() function takes an optional filter
list which can be used to return strains associated with a
particular species
}
\details{
See http://fsis.glfc.com/api/public_urls/#v1-common-strains-list
for the full list of available filter keys (query parameters)
}
\examples{

get_strains()
get_strains(list(species="LAT"))
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
