% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_filters.R
\name{refresh_filters}
\alias{refresh_filters}
\title{Refresh the global api filter list}
\usage{
refresh_filters(endpoint)
}
\arguments{
\item{endpoint}{\itemize{
\item the name of the api endpoint
}}
}
\value{
list
}
\description{
A function to refresh or fetch available filters for an endpoint.
}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
