% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_filters.R
\name{show_filters}
\alias{show_filters}
\title{Show filters available on the provided api endpoint}
\usage{
show_filters(endpoint = "", filter_like = "")
}
\arguments{
\item{endpoint}{\itemize{
\item the name of the api endpoint. If no name is
provided, a list of available endpoints will be displayed.
}}

\item{filter_like}{\itemize{
\item a partial string to match filter against
}}
}
\description{
Many of the  GLFSD API endpoints have a number of
available filters that allows the response to be taylored to your
particular needs.  This function takes and api endpoint name and
prints out all of the available filters.  If 'filter_like' is
provided, only those filters that contain the provided string are
printed.  If no endpoint name is provided, a list of available
endpoints will be displayed.
}
\details{
See http://fsis.glfc.org/api/public_urls/ for more information on the
available filter keys (query parameters) for each endpoint
}
\examples{
show_filters()
show_filters("strain")
show_filters("events", filter_like = "clip")

}
\author{
Adam Cottrill \email{adam.cottrill@ontario.ca}
}
